requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pywin32>=306; platform_system == "Windows"
xlrd>=2.0.1
Pillow>=10.0.0